requirements.txt
fastmcp>=2.13.0.1
requests>=2.31.0
python-dotenv>=1.2.1